myproject/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.7.0
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
